+++ transport/internal/errors/peers.go
@@ -54,3 +54,23 @@ type ErrPeerListNotStarted string
 func (e ErrPeerListNotStarted) Error() string {
 	return fmt.Sprintf("%s has not been started or was stopped", string(e))
 }
+
+// ErrInvalidPeerConversion is called when a peer can't be properly converted
+type ErrInvalidPeerConversion struct {
+	Peer         transport.Peer
+	ExpectedType string
+}
+
+func (e ErrInvalidPeerConversion) Error() string {
+	return fmt.Sprintf("cannot convert peer (%v) to type %s", e.Peer, e.ExpectedType)
+}
+
+// ErrInvalidAgentConversion is called when an agent can't be properly converted
+type ErrInvalidAgentConversion struct {
+	Agent        transport.PeerAgent
+	ExpectedType string
+}
+
+func (e ErrInvalidAgentConversion) Error() string {
+	return fmt.Sprintf("cannot convert agent (%v) to type %s", e.Agent, e.ExpectedType)
+}
